Create a new simulation object

Create a new simulation object. This is typically the first function to be called when running a simulation using SimEngine. Most other SimEngine functions take a simulation object as their first argument.

Usage

new_sim()

Value

A simulation object, of class sim_obj

See also

Examples

sim <- new_sim()
print(sim)